+++ tests/special/dll2klib-e2e/invalid-pointer-function.kt
@@ -1,0 +1,5 @@
+package invalidpointerfunction
+
+import kotlin.clr.ClrPointer
+
+fun invalid(value: (ClrPointer<Int>) -> Unit): Unit = Unit